requirements.txt
altair
pandas
streamlit
streamlit_authenticator
deta
python-dotenv